make main.js load compiled code
